use --only-show-errors to call CLI commands if user specify

+++ src/image-copy/azext_imagecopy/custom.py
@@ -21,11 +21,13 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
               target_resource_group_name, temporary_resource_group_name='image-copy-rg',
               source_type='image', cleanup=False, parallel_degree=-1, tags=None, target_name=None,
               target_subscription=None, export_as_snapshot='false', timeout=3600):
+    only_show_errors = cmd.cli_ctx.only_show_errors
     if cleanup:
         # If --cleanup is set, forbid using an existing temporary resource group name.
         # It is dangerous to clean up an existing resource group.
         cli_cmd = prepare_cli_command(['group', 'exists', '-n', temporary_resource_group_name],
-                                      output_as_json=False)
+                                      output_as_json=False,
+                                      only_show_errors=only_show_errors)
         cmd_output = run_cli_command(cli_cmd)
         if 'true' in cmd_output:
             raise CLIError('Don\'t specify an existing resource group in --temporary-resource-group-name '
@@ -35,7 +37,8 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
     logger.warning("Getting OS disk ID of the source VM/image")
     cli_cmd = prepare_cli_command([source_type, 'show',
                                    '--name', source_object_name,
-                                   '--resource-group', source_resource_group_name])
+                                   '--resource-group', source_resource_group_name],
+                                  only_show_errors=only_show_errors)
 
     json_cmd_output = run_cli_command(cli_cmd, return_as_json=True)
 
@@ -96,14 +99,16 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
                                        '--resource-group', source_resource_group_name,
                                        '--source', source_os_disk_id,
                                        '--source-storage-account-id', source_storage_account_id,
-                                       '--hyper-v-generation', hyper_v_generation])
+                                       '--hyper-v-generation', hyper_v_generation],
+                                      only_show_errors=only_show_errors)
     else:
         cli_cmd = prepare_cli_command(['snapshot', 'create',
                                        '--name', source_os_disk_snapshot_name,
                                        '--location', snapshot_location,
                                        '--resource-group', source_resource_group_name,
                                        '--source', source_os_disk_id,
-                                       '--hyper-v-generation', hyper_v_generation])
+                                       '--hyper-v-generation', hyper_v_generation],
+                                      only_show_errors=only_show_errors)
 
     run_cli_command(cli_cmd)
 
@@ -117,7 +122,8 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
     cli_cmd = prepare_cli_command(['snapshot', 'grant-access',
                                    '--name', source_os_disk_snapshot_name,
                                    '--resource-group', source_resource_group_name,
-                                   '--duration-in-seconds', str(timeout)])
+                                   '--duration-in-seconds', str(timeout)],
+                                  only_show_errors=only_show_errors)
 
     json_output = run_cli_command(cli_cmd, return_as_json=True)
 
@@ -132,15 +138,18 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
     transient_resource_group_location = target_location[0].strip()
     create_resource_group(transient_resource_group_name,
                           transient_resource_group_location,
-                          target_subscription)
+                          target_subscription,
+                          only_show_errors)
 
     target_locations_count = len(target_location)
     logger.warning("Target location count: %s", target_locations_count)
 
     create_resource_group(target_resource_group_name,
                           target_location[0].strip(),
-                          target_subscription)
+                          target_subscription,
+                          only_show_errors)
 
+    pool = None
     try:
 
         # try to get a handle on arm's 409s
@@ -159,7 +168,7 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
                                     source_object_name, source_os_disk_snapshot_name, source_os_disk_snapshot_url,
                                     source_os_type, target_resource_group_name, azure_pool_frequency,
                                     tags, target_name, target_subscription, export_as_snapshot, timeout,
-                                    hyper_v_generation)
+                                    hyper_v_generation, only_show_errors=only_show_errors)
         else:
             if parallel_degree == -1:
                 pool = Pool(target_locations_count)
@@ -174,7 +183,7 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
                               source_object_name, source_os_disk_snapshot_name, source_os_disk_snapshot_url,
                               source_os_type, target_resource_group_name, azure_pool_frequency,
                               tags, target_name, target_subscription, export_as_snapshot, timeout,
-                              hyper_v_generation))
+                              hyper_v_generation, only_show_errors))
 
             logger.warning("Starting async process for all locations")
 
@@ -189,7 +198,8 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
         if cleanup:
             logger.warning('To cleanup temporary resources look for ones tagged with "image-copy-extension". \n'
                            'You can use the following command: az resource list --tag created_by=image-copy-extension')
-        pool.terminate()
+        if pool is not None:
+            pool.terminate()
         return
 
     # Cleanup
@@ -199,29 +209,35 @@ def imagecopy(cmd, source_resource_group_name, source_object_name, target_locati
         # Delete resource group
         cli_cmd = prepare_cli_command(['group', 'delete', '--no-wait', '--yes',
                                        '--name', transient_resource_group_name],
-                                      subscription=target_subscription)
+                                      subscription=target_subscription,
+                                      only_show_errors=only_show_errors)
         run_cli_command(cli_cmd)
 
         # Revoke sas for source snapshot
         cli_cmd = prepare_cli_command(['snapshot', 'revoke-access',
                                        '--name', source_os_disk_snapshot_name,
-                                       '--resource-group', source_resource_group_name])
+                                       '--resource-group', source_resource_group_name],
+                                      only_show_errors=only_show_errors)
         run_cli_command(cli_cmd)
 
         # Delete source snapshot
         # TODO: skip this if source is snapshot and not creating a new one
         cli_cmd = prepare_cli_command(['snapshot', 'delete',
                                        '--name', source_os_disk_snapshot_name,
-                                       '--resource-group', source_resource_group_name])
+                                       '--resource-group', source_resource_group_name],
+                                      only_show_errors=only_show_errors)
         run_cli_command(cli_cmd)
 
+    logger.warning('Image copy finished')
 
-def create_resource_group(resource_group_name, location, subscription=None):
+
+def create_resource_group(resource_group_name, location, subscription=None, only_show_errors=None):
     # check if target resource group exists
     cli_cmd = prepare_cli_command(['group', 'exists',
                                    '--name', resource_group_name],
                                   output_as_json=False,
-                                  subscription=subscription)
+                                  subscription=subscription,
+                                  only_show_errors=only_show_errors)
 
     cmd_output = run_cli_command(cli_cmd)
 
@@ -233,6 +249,7 @@ def create_resource_group(resource_group_name, location, subscription=None):
     cli_cmd = prepare_cli_command(['group', 'create',
                                    '--name', resource_group_name,
                                    '--location', location],
-                                  subscription=subscription)
+                                  subscription=subscription,
+                                  only_show_errors=only_show_errors)
 
     run_cli_command(cli_cmd)
